move handlers to message context
