+++ src/smart-components/NotificationChannel/_externals/react-message-template-view/src/context/MessageContextProvider.tsx
@@ -1,17 +1,20 @@
 // create a context provider for MessageComponent
 //  - this is a wrapper for MessageComponent
-import React from 'react';
+import React, { Children } from 'react';
 
 import { BaseMessage } from '@sendbird/chat/message';
+import { Action } from '../../../message-template/src';
+
 
 export type MessageContextProps = {
   message: BaseMessage;
   children?: React.ReactElement;
+  handleWebAction?(event: React.SyntheticEvent, action: Action, message: BaseMessage);
+  handleCustomAction?(event: React.SyntheticEvent, action: Action, message: BaseMessage);
+  handlePredefinedAction?(event: React.SyntheticEvent, action: Action, message: BaseMessage);
 };
 
-export type MessageContextInterface = {
-  message: BaseMessage;
-}
+export type MessageContextInterface = Omit<MessageContextProps, 'children'>;
 
 const MessageContext = React.createContext<MessageContextProps | null>(undefined);
 const MessageProvider: React.FC<MessageContextProps> = (
@@ -19,11 +22,17 @@ const MessageProvider: React.FC<MessageContextProps> = (
 ) => {
   const {
     message,
+    handleWebAction,
+    handleCustomAction,
+    handlePredefinedAction,
     children,
   } = props;
 
   const value = React.useMemo(() => ({
     message,
+    handleWebAction,
+    handleCustomAction,
+    handlePredefinedAction,
   }), [message?.updatedAt]);
 
   return (
