+++ src/smart-components/NotificationChannel/context/dux/initialState.ts
@@ -1,6 +1,5 @@
 import type { GroupChannel } from "@sendbird/chat/groupChannel";
 import type { BaseMessage, MessageListParams } from "@sendbird/chat/message";
-import { number } from "prop-types";
 
 export declare type NotficationChannelStateInterface = {
   uiState: 'loading' | 'initialized' | 'invalid';
